% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikehamilton.R
\docType{package}
\name{bikehamilton}
\alias{bikehamilton}
\alias{bikehamilton-package}
\title{bikehamilton: This package features different cycling routes from points of interest
in Hamilton, Ontario based on level of comfort.}
\source{
University of Toronto, Data Management Group {http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}
}
\description{
Routes are categorized as fast (most contact with traffic),
balanced (some contact with traffic), and quiet (minimal contact with traffic).
The package includes a sample function and
a sample dataset with their respective documentation.
}
\author{
Elise Desjardins, Faculty of Health Sciences, McMaster University \email{desjae@mcmaster.ca}
}
\keyword{Canada}
\keyword{bicycling,}
\keyword{routing,}
\keyword{urban,}
